Don't start and enable livvirtd
libvirtd is now automagically started when virt-manager connect to it,
so there's no need to start libvirtd in advance. Also, libvirtd
automatically stops if no clients are connected and no domains are
running for a while.

This commit should've been included in the initial import, but was
missed. See:
Ichimonji10/impedimenta@328f4638414e41c00e17eacccd9869e30bf0eeb5