Make routers nftables.conf more legible
Most changes are minimal, like re-ordering blocks and fiddling with
whitespace. The only especially interesting change is the rename of
several tables:

* input_* → input_from_*
* forward_* → forward_from_*

This change is designed to make it more clear that these tables deal
with packets coming in from * or being forwarded from *, and it adds
symmetry to the table named forward_to_dmz.