Deduplicate definitions in routers nftables.conf
Define a variable listing all forwarded TCP ports, and interpolate that
variable as needed. A set can't be used for this purpose, as it can only
be used within the same table that it's defined, and the set of
forwarded ports is referenced once from the filter table and once from
the nat table.