Totally revamp networking configuration
Create several new roles:

*   networking-nm-branch
*   networking-nm-leaf
*   networking-sd-leaf
*   nftables-branch
*   nftables-leaf

The `networking-nm-*` roles manage networking with NetworkManager and
related tools, whereas the `networking-sd-*` role manages networking
with systemd and related tools. A "branch" host forward packets between
a wireless interface and a wired interface. A "leaf" host is at the
network edge and doesn't forward packets.

Drop the following roles:

*   dns
*   ip-forwarding
*   netctl-routers (netctl-generic is already gone)
*   nftables-generic
*   nftables-routers
*   systemd-networkd-generic

As a result of these changes, there are no more "routers" hosts, and
pine is not accessible from the internet. But this change at least gets
all hosts back online, and the inaccessibility of pine in particular can
be worked out as time goes on.