Drop custom-repo role
It's been broken for a while. Best to drop it and pick it up later as
needed, as compared to carting around broken code that makes maintenance
more difficult.

This repository provided several packages that can no longer be
installed but that other roles reference. This commit uninstalls or
removes references to several of those packages, such as bandwidth or
vim-badwolf-git, respectively. Other references, such as to airsonic,
are left in place — but those roles should be disabled.