Drop bittorrent_hosts group
If bittorrent is to be run, it will be through Docker. Think, say, of a
container that makes a wireshark connection to a VPN provider, and a
second container that sends all traffic through the first. In this case,
there's no need for the host OS to mangle its firewall on behalf of the
bittorrent application.