Merge all group vars into a single file
Doing this makes it easier to manage variables. There's no case where a
variable is defined differently based on which group a host is in. This
does have a potential security implication, in that a task that should
only execute on a workstation might execute on a server, now that the
secret is available to it.