Only call converted roles from site.yml playbook
Several roles still need to be converted, but they are so few, and the
need to finally convert workstations to Fedora great enough, that it's
sensible to simply list the few roles that still haven't been rewritten
as commented-out items in the main playbook. This makes it very clear
which bits of configuration are still lacking.